+++ package.json
@@ -3,7 +3,7 @@
     "version": "0.1.0",
     "private": true,
     "scripts": {
-        "dev": "next dev",
+        "dev": "next dev --turbopack",
         "build": "next build",
         "start": "next start",
         "lint": "next lint"
@@ -53,15 +53,15 @@
         "zod": "^3.24.1"
     },
     "devDependencies": {
-        "@types/node": "^20.17.19",
-        "@types/react": "^18.2.6",
-        "@types/react-dom": "^18.2.4",
         "autoprefixer": "^10.4.14",
         "eslint": "^8.40.0",
         "eslint-config-next": "^13.4.2",
         "postcss": "^8.4.23",
         "tailwindcss": "^3.3.2",
-        "typescript": "^5.0.4"
+        "typescript": "^5.0.4",
+        "@types/node": "^20",
+        "@types/react": "^19",
+        "@types/react-dom": "^19"
     },
     "browser": {
         "net": false
