--- pages/agent/packag.json
@@ -1,35 +1,0 @@
-{
-    "name": "@goat-sdk/plugin-balmy",
-    "version": "0.1.11",
-    "files": ["dist/**/*", "README.md", "package.json"],
-    "scripts": {
-        "build": "tsup",
-        "clean": "rm -rf dist",
-        "test": "vitest run --passWithNoTests"
-    },
-    "sideEffects": false,
-    "main": "./dist/index.js",
-    "module": "./dist/index.mjs",
-    "types": "./dist/index.d.ts",
-    "dependencies": {
-        "@balmy/sdk": "0.6.1",
-        "@goat-sdk/core": "workspace:*",
-        "@goat-sdk/wallet-evm": "workspace:*",
-        "viem": "catalog:",
-        "zod": "catalog:"
-    },
-    "peerDependencies": {
-        "@goat-sdk/core": "workspace:*",
-        "viem": "catalog:"
-    },
-    "homepage": "https://ohmygoat.dev",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/goat-sdk/goat.git"
-    },
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/goat-sdk/goat/issues"
-    },
-    "keywords": ["ai", "agents", "web3"]
-}
